Show play or pause icon depends on state

+++ kartoffel-music-app/Sources/CommonModels/PlayState.swift
@@ -1,5 +1,5 @@
 
 public enum PlayState: Equatable, Hashable {
     case stop
-    case playing(progress: Int)
+    case playing(progress: Double)
 }
